create new file to move logic and jsx to Notification.jsx
